pyproject.toml
[tool.poetry]
name = "hyperdeck-timeremaining"
version = "0.1.0"
description = "Displays the remaining time of hyperdeck media players"
authors = ["Ulrich Sommer <mail@ulrich-sommer.de"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
Flask-SocketIO = "^5.3.2"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
